+++ backend/CalculadoraCalorias.Core/Domain/InternalDTO/RefeicaoDTO.cs
@@ -9,7 +9,7 @@ namespace CalculadoraCalorias.Core.Domain.InternalDTO
     {
         public long Id { get; set; }
         public TipoRefeicaoEnum TipoRefeicao { get; set; }
-        public double? Calorias {  get; set; }
+        public int? Calorias {  get; set; }
         public DateOnly Data { get; set; }
     }
 }
